% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_chars.R
\name{compress_chars}
\alias{compress_chars}
\alias{uncompress_chars}
\title{Recursively Call \code{\link{char_map}}/\code{\link{map2char}} on an Object}
\usage{
compress_chars(
  l,
  limit = 0L,
  compress_altreps = c("if_allocated", "yes", "no")
)

uncompress_chars(l)
}
\arguments{
\item{l}{an object, typically a list or an environment}

\item{limit}{the minimum length of a character vector for
\code{\link{char_map}} to be applied}

\item{compress_altreps}{should a character vector be compressed if it is an
ALTREP? The default "if_allocated" only does so if the regular
representation was already created. This was chosen as the default because
in this case is is the regular representation which would be serialized.}
}
\description{
Recursively Call \code{\link{char_map}}/\code{\link{map2char}} on an Object
}
\note{
The object returned by \code{compress_chars} might be an invalid S3
  object, e.g. if \code{l} is a data frame. These functions are intended to
  be called immediately before and after (de)serializing the object, i.e.
  compress -> serialize -> store/transfer -> de-serialize -> uncompress.
}
