% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_cpu_pool}
\alias{create_cpu_pool}
\alias{destroy_cpu_pool}
\title{CPU Pools}
\usage{
create_cpu_pool(ncpu = parallel::detectCores())

destroy_cpu_pool(p)
}
\arguments{
\item{ncpu}{the number of CPUs in the pool.}

\item{p}{an object as returned by \code{create_cpu_pool}.}
}
\description{
Create and destroy CPU pools to be passed to \code{\link{mclapply}} as
argument \code{mc.cpu.pool}.
}
