% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_reset_trigger.R
\name{gc_reset_trigger}
\alias{gc_reset_trigger}
\title{Reset the Trigger Thresholds of the Garbage Collector}
\usage{
gc_reset_trigger(
  print_stats = TRUE,
  verbose = getOption("verbose"),
  reset = TRUE
)
}
\arguments{
\item{print_stats}{should the memory statistics, as returned by
\code{\link{gc}()}, be printed once at the beginning and once at the end of
this function.}

\item{verbose, reset}{arguments passed to \code{\link{gc}()}.}
}
\value{
the return value of the final call to \code{\link{gc}()}.
}
\description{
This function repeatedly triggers full garbage collections until the trigger
thresholds no longer decrease.
}
\details{
After removing huge objects, the "gc trigger" thresholds will remain
  on a high level for a very long time because they are only gradually
  adjusted downwards on full collections, which are very rare. Higher
  thresholds imply a high memory usage of the R process because collections
  are triggered less often. This is a particular problem when using
  fork-based parallelization, because the child processes will "inherit the
  wasteful attitude" from their parent, such that the whole process family
  might end up using unbearably much memory.
}
\note{
Trigger thresholds can also increase due to calling this function. This
  is not a bug but due to the internals of the garbage collector.
}
